DEV: misc fixes

+++ workflow/scripts/copyOrGzip.py
@@ -1,5 +1,5 @@
-import gzip
-import shutil
+import re
+import os
 
 
 def gzip_file(input_file, output_file):
@@ -10,13 +10,11 @@ def gzip_file(input_file, output_file):
         input_file (str): filepath of input file (gzipped or not)
         output_file (str): filepath of output gzipped file
     """
-    is_gzipped = input_file.endswith('.gz')
-    if is_gzipped:
-        shutil.copyfile(input_file, output_file)
+    fasta_regex = re.compile(r"^[a-zA-Z0-9_-]+\.(fa|fasta|fna|ffn|faa|frn)(\.gz)?$")
+    if fasta_regex.match(input_file):
+        os.system(f"seqtk -F I {input_file} | gzip -1 -c > {output_file}")
     else:
-        with open(input_file, 'rb') as f_in:
-            with gzip.open(output_file, 'wb') as f_out:
-                shutil.copyfileobj(f_in, f_out)
+        os.system(f"seqtk {input_file} | gzip -1 -c > {output_file}")
 
 
 def main(**kwargs):
@@ -24,7 +22,10 @@ def main(**kwargs):
     if "input_r2" in kwargs.keys():
         gzip_file(kwargs["input_r2"], kwargs["output_r2"])
     if "input_s" in kwargs.keys():
-        open(kwargs["output_s"], "w").close()
+        if os.path.exists(kwargs["input_s"]) and os.path.getsize(kwargs["input_s"]) > 0:
+            gzip_file(kwargs["input_s"], kwargs["output_s"])
+        else:
+            open(kwargs["output_s"], "w").close()
 
 
 if __name__ == "__main__":
